Completando semana 9 de JAVA
