+++ Java/Semana 9 (29-05 al 02-06)/ManejoDeExcepciones1/nbproject/private/private.properties
@@ -5,4 +5,4 @@ do.jlink=false
 javac.debug=true
 javadoc.preview=true
 jlink.strip=false
-user.properties.file=C:\\Users\\Usuario\\AppData\\Roaming\\NetBeans\\12.1\\build.properties
+user.properties.file=C:\\Users\\lucas\\AppData\\Roaming\\NetBeans\\12.1\\build.properties
