+++ Java/Semana 9 (29-05 al 02-06)/ManejoDeExcepciones1/src/Excepciones/OperacionExcepcion.java
@@ -1,0 +1,12 @@
+
+package Excepciones;
+
+
+public class OperacionExcepcion extends RuntimeException{
+    //Constructor
+    public OperacionExcepcion(String mensaje){
+        super(mensaje);
+        
+    }
+    
+}
