--- Java/Semana 9 (29-05 al 02-06)/ManejoDeExcepciones1/src/Excepciones/OperaciónException.java
@@ -1,12 +1,0 @@
-
-package Excepciones;
-
-
-public class OperaciónException extends RuntimeException{
-    //Constructor
-    public OperaciónException(String mensaje){
-        super(mensaje);
-        
-    }
-    
-}
