+++ Java/Semana 9 (29-05 al 02-06)/ManejoDeExcepciones1/src/aritmetica/Aritmetica.java
@@ -1,0 +1,20 @@
+
+package aritmetica;
+
+import Excepciones.OperacionExcepcion;
+
+
+public class Aritmetica {
+    public static int division(int numerador, int denominador){
+        if (denominador == 0){
+            throw new OperacionExcepcion("División entre cero");  
+        }
+        return numerador / denominador;
+    }
+    
+     public static class division {
+
+        public division() {
+        }
+    }
+}
