--- Java/Semana 9 (29-05 al 02-06)/ManejoDeExcepciones1/src/aritmetica/Aritmética.java
@@ -1,18 +1,0 @@
-
-package aritmetica;
-
-
-public class Aritmética {
-    public static int division(int númerador, int denominador){
-        if (denominador == 0){
-            throw new OperacionException("División entre cero");  
-        }
-        return numerador / denominador;
-    }
-
-    public static class division {
-
-        public division() {
-        }
-    }
-}
