+++ Java/Semana 9 (29-05 al 02-06)/ManejoDeExcepciones1/src/test/TestExcepciones.java
@@ -1,18 +1,17 @@
-
 package test;
 
-import aritmetica.Aritmética.division;
-
-
+import static aritmetica.Aritmetica.division;
+import Excepciones.OperacionExcepcion;
 
 public class TestExcepciones {
     public static void main(String[] args) {
         int resultado = 0;
         try{
-            resultado = Aritmética.division(10, 2);
-        }catch(OperaciónException e){
+            resultado = division(10, 2);
+            
+        }catch(OperacionExcepcion e){
             System.out.println("Ocurrió un error de tipo OperaciónExcepción");
-            System.out.println(e.getmessage());
+            System.out.println(e.getMessage());
         } catch(Exception e){
             System.out.println("Ocurrió un error");
             e.printStackTrace(System.out);
